+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/document/DocumentService.java
@@ -47,7 +47,7 @@ public interface DocumentService
      * @param document the Document to save.
      * @return the generated identifier.
      */
-    int saveDocument( Document document );
+    long saveDocument( Document document );
 
     /**
      * Retrieves the Document with the given identifier.
@@ -55,7 +55,7 @@ public interface DocumentService
      * @param id the identifier of the Document.
      * @return the Document.
      */
-    Document getDocument( int id );
+    Document getDocument( long id );
 
     /**
      * Retrieves the Document with the given identifier.
